Add ability to config pointfree command
It looks like when using ! command in (n)vim, it will run a command in
new shell, which will reset PATH base in shell setup. So if you use
direnv (or similar tool) to add pointfree into your path, the current
way to trigger pointfree won't work.

This fix introduce a new g:potion_command to allow user to override it,
and use exepath to find pointfree based on current shell session before
calling it